Move classes and tests to PaymentAuthConfig
